+++ internal/utils/envparse.go
@@ -24,7 +24,7 @@ type Version struct {
 	Additional string
 }
 
-var variables *config.Variables
+var cfg *config.Config
 
 func (e *ParseError) Error() string {
 	if e.Line > 0 {
@@ -44,26 +44,38 @@ func (v *Version) String() string {
 	// 1. construct base version string
 	base := fmt.Sprintf("%d.%d.%d", v.Major, v.Minor, v.Patch)
 	// split MergeRequestTargetBranchName on /
-	targetBranch := strings.Split(strings.ToLower(variables.MergeRequestTargetBranchName), "/")[0]
-	// log.Printf("Target branch: %s\n", targetBranch)
+	targetBranch := strings.Split(strings.ToLower(cfg.Variables.MergeRequestTargetBranchName), "/")
+	// TODO: handle error when unable to split?
+	var tb string
+	if len(targetBranch) == 1 {
+		tb = targetBranch[0]
+	} else if len(targetBranch) == 2 {
+		tb = fmt.Sprintf("%s-%s", targetBranch[0], targetBranch[1])
+	}
+	if cfg.Debug {
+		log.Printf("targetBranch: %s\n", targetBranch)
+	}
 	// 2. if on defalt or release branch return version string
-	if targetBranch == variables.DefaultBranch || targetBranch == "release" {
+	if tb == cfg.Variables.DefaultBranch || strings.Contains(tb, "release") {
 		return fmt.Sprintf("v%s", base)
 	} else {
 		// 3. if target branch is not default branch, append branch to base
-		base = fmt.Sprintf("%s-%s", base, targetBranch)
+		base = fmt.Sprintf("%s-%s", base, tb)
 	}
 	// 4. check if we need to append additional options
 	if v.Additional != "" {
 		base = fmt.Sprintf("%s-%s", base, v.Additional)
 	}
 	// 5. Add build number
-	return fmt.Sprintf("%s+%d", base, variables.PipelineIid)
+	return fmt.Sprintf("%s+%d", base, cfg.Variables.PipelineIid)
 }
 
-func GetVersion(v *config.Variables) string {
+func GetVersion(c *config.Config) string {
 	// set variables
-	variables = v
+	cfg = c
+	if cfg.Debug {
+		log.Println("GetVersion | Setting variables")
+	}
 	// version is in the format of vX.Y.Z
 	// we want to return X.Y.Z (and optionally -ADDOPTS)
 	env := make(map[string]string)
@@ -90,7 +102,6 @@ func GetVersion(v *config.Variables) string {
 		i++
 		k, v, err := parseLine(scanner.Bytes())
 		if err != nil {
-			log.Println(parseError(i, err))
 			return vr.String()
 		}
 
@@ -134,8 +145,7 @@ func parseLine(line []byte) ([]byte, []byte, error) {
 		return nil, nil, fmt.Errorf("no equals sign found")
 	}
 
-	// Split the line into two parts
-	// split line by =
+	// Split the line into two parts by = sign
 	kv := bytes.Split(line, []byte("="))
 	k := bytes.TrimSpace(kv[0])
 	v := bytes.TrimSpace(kv[1])